code cleanup, x64 fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@11163 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Clist_nicer/src/clui.cpp
@@ -1234,7 +1234,7 @@ skipbg:
 	}
 
 	case WM_TIMER:
-		if ((int)wParam == TM_AUTOALPHA) {
+		if (wParam == TM_AUTOALPHA) {
 			int inwnd;
 
 			if (GetForegroundWindow() == hwnd) {
